[util] Organize and package `make_new_dif.py` classes.
IP HJSON file are parsed by the `util/make_new_dif.py` tool to load
IP metadata needed for auto-generating DIFs. This python code was
originally all contained in the `util/make_new_dif.py` script. By
repackaging it, we can reuse this code in the
`util/check_dif_statuses.py` script.
