+++ util/make_new_dif/templates/dif_autogen.h.tpl
