+++ util/make_new_dif/templates/dif_template.h.tpl
